+++ tests/framework/results/test_observer.py
@@ -1,29 +1,31 @@
 import pytest
 from layered_config_tree.main import LayeredConfigTree
+from pytest_mock import MockerFixture
 
+from vivarium.framework.engine import Builder
 from vivarium.framework.results.observer import Observer
 
 
 class TestObserver(Observer):
-    def register_observations(self, builder):
+    def register_observations(self, builder: Builder) -> None:
         pass
 
 
 class TestDefaultObserverStratifications(Observer):
-    def register_observations(self, builder):
+    def register_observations(self, builder: Builder) -> None:
         pass
 
 
 class TestObserverStratifications(Observer):
-    def register_observations(self, builder):
+    def register_observations(self, builder: Builder) -> None:
         pass
 
     @property
-    def configuration_defaults(self):
+    def configuration_defaults(self) -> dict[str, str]:
         return {"foo": "bar"}
 
 
-def test_observer_instantiation():
+def test_observer_instantiation() -> None:
     observer = TestObserver()
     assert observer.name == "test_observer"
 
@@ -35,7 +37,9 @@ def test_observer_instantiation():
         (True, None),
     ],
 )
-def test_set_results_dir(is_interactive, results_dir, mocker):
+def test_set_results_dir(
+    is_interactive: bool, results_dir: str | None, mocker: MockerFixture
+) -> None:
     builder = mocker.Mock()
     if is_interactive:
         builder.configuration = LayeredConfigTree()