+++ CHANGELOG.rst
@@ -1,3 +1,9 @@
+**3.3.13 - 03/21/25**
+
+  - Type-hinting: Fix mypy errors in tests/framework/results/test_observer.py
+  - Type-hinting: Fix mypy errors in tests/framework/results/test_stratificaton.py
+  - Type-hinting: Fix mypy errors in tests/framework/results/test_pbservation.py
+
 **3.3.12 - 03/17/25**
 
   - Type-hinting: Fix mypy errors in tests/framework/test_plugins.py